refactor(capabilities): nest contentgen under a shared submodule
Move `src/contentgen/` → `src/shared/contentgen/` to make its
not-a-capability role explicit in the module tree. Introduce
`src/shared/mod.rs` as the container for cross-capability infrastructure.
Rewrite every in-crate `crate::contentgen::` path to
`crate::shared::contentgen::`. The flattened public re-export surface in
`lib.rs` (`AnthropicAdapter`, `GeminiAdapter`, `StubAnthropicAdapter`,
`StubGeminiAdapter`, `TaskQueue`, `stage_gen_output`) is byte-for-byte
unchanged — external callers are unaffected.

Closes #2267.
