+++ crates/aether-capabilities/src/gemini/adapter.rs
@@ -9,11 +9,11 @@ use std::time::Duration;
 
 use serde_json::Value;
 
-use crate::contentgen::adapter::{
+use crate::shared::contentgen::adapter::{
     AdapterUsage, GeminiAdapter, GeminiArtifact, GeminiImageRequest, GeminiMusicRequest,
     GeminiResponse,
 };
-use crate::contentgen::shared;
+use crate::shared::contentgen::shared;
 
 use super::{AspectRatio, GeminiError, ImageSize, ThinkingLevel};
 use super::{error, lyria, nanobanana};
@@ -271,7 +271,7 @@ mod tests {
     /// `includeThoughts`, and `tools[0].google_search` (issue 1167).
     #[test]
     fn nanobanana_body_carries_set_params() {
-        use crate::contentgen::adapter::GeminiImageRequest;
+        use crate::shared::contentgen::adapter::GeminiImageRequest;
         let body = super::build_nanobanana_body(&GeminiImageRequest {
             model: "gemini-3.1-flash-image-preview".to_string(),
             prompt: "a cat".to_string(),
@@ -295,7 +295,7 @@ mod tests {
     /// `aspectRatio` survives under `imageConfig`.
     #[test]
     fn nanobanana_body_omits_unset_params() {
-        use crate::contentgen::adapter::GeminiImageRequest;
+        use crate::shared::contentgen::adapter::GeminiImageRequest;
         let body = super::build_nanobanana_body(&GeminiImageRequest {
             model: "gemini-3.1-flash-image-preview".to_string(),
             prompt: "a cat".to_string(),
@@ -317,7 +317,7 @@ mod tests {
     #[ignore = "needs GEMINI_API_KEY"]
     fn gemini_nanobanana_smoke() {
         use super::UreqGeminiAdapter;
-        use crate::contentgen::adapter::{GeminiAdapter, GeminiImageRequest};
+        use crate::shared::contentgen::adapter::{GeminiAdapter, GeminiImageRequest};
         use std::env;
         use std::time::Duration;
         // Test-only: the live-API smoke reads an external credential
@@ -342,7 +342,7 @@ mod tests {
     #[ignore = "needs GEMINI_API_KEY"]
     fn gemini_lyria_smoke() {
         use super::UreqGeminiAdapter;
-        use crate::contentgen::adapter::{GeminiAdapter, GeminiMusicRequest};
+        use crate::shared::contentgen::adapter::{GeminiAdapter, GeminiMusicRequest};
         use std::env;
         use std::time::Duration;
         // Test-only: the live-API smoke reads an external credential
