+++ crates/aether-capabilities/src/gemini/mod.rs
@@ -5,7 +5,7 @@
 //!
 //! Like `aether.anthropic` (issue 1014), long-tail calls ride the
 //! ADR-0093 hold-until-resolve dispatch: the generate handler submits the
-//! blocking `ureq` call to a [`TaskQueue`](crate::contentgen::TaskQueue)
+//! blocking `ureq` call to a [`TaskQueue`](crate::shared::contentgen::TaskQueue)
 //! (`ctx.dispatch_blocking`), so an ephemeral worker runs the call and
 //! stages the generated bytes to `save://gen/<uuid>.{png,wav}` (binary
 //! outputs never ride the mail wire); the completion routes back to the
