+++ crates/aether-capabilities/src/shared/contentgen/mod.rs
