+++ crates/aether-capabilities/src/shared/contentgen/shared.rs
