+++ crates/aether-capabilities/src/shared/contentgen/staging.rs
