+++ crates/aether-capabilities/src/shared/mod.rs
@@ -1,0 +1,3 @@
+//! Shared infrastructure consumed by multiple capabilities but not itself a capability.
+#[cfg(not(target_arch = "wasm32"))]
+pub mod contentgen;
